Created a volume post processor and a spotlight
I made a global volume post processor with red lighting to it to create the idea of romance and danger. I also added a little more bloom to it especially since it's supposed to be in a darkish restaurant so that you could still see the props petty well. I also edited the spotlights that were already there and made it so that they were the only ones that created a shadow.